BrickAI/requirements.txt
# Core
Flask==3.0.3
Werkzeug==3.0.4
python-dotenv==1.0.1
bleach==6.2.0
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
gunicorn==22.0.0

# Database
pymongo==4.10.1
bcrypt==4.2.1

# Images / ML
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
torch>=2.4,<2.6
torchvision>=0.19,<0.21

# LLM
openai>=1.109.0,<2.0.0

# Azure Storage
azure-storage-blob==12.24.0
azure-identity==1.19.0

# Optional (defects)
ultralytics==8.3.0
